refactor(server): create demos.cpp
containing all recording functions for the server

TODO: test demo playback (is it really working already on a remote server?)
